rename: rename directory (pallete->palette)
