+++ src/social-service/src/main/java/org/palette/easelsocialservice/config/Neo4jConfig.java
@@ -1,4 +1,4 @@
-package org.pallete.easelsocialservice.config;
+package org.palette.easelsocialservice.config;
 
 import org.springframework.context.annotation.Bean;
 import org.neo4j.cypherdsl.core.renderer.Configuration;
