+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/LinkRepository.java
@@ -1,6 +1,6 @@
-package org.pallete.easelsocialservice.persistence;
+package org.palette.easelsocialservice.persistence;
 
-import org.pallete.easelsocialservice.persistence.domain.Link;
+import org.palette.easelsocialservice.persistence.domain.Link;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 public interface LinkRepository extends Neo4jRepository<Link, String> {
