+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/MediaRepository.java
@@ -1,6 +1,6 @@
-package org.pallete.easelsocialservice.persistence;
+package org.palette.easelsocialservice.persistence;
 
-import org.pallete.easelsocialservice.persistence.domain.Media;
+import org.palette.easelsocialservice.persistence.domain.Media;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 public interface MediaRepository extends Neo4jRepository<Media, String> {
