+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/PaintRepository.java
@@ -1,6 +1,6 @@
-package org.pallete.easelsocialservice.persistence;
+package org.palette.easelsocialservice.persistence;
 
-import org.pallete.easelsocialservice.persistence.domain.Paint;
+import org.palette.easelsocialservice.persistence.domain.Paint;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 import java.util.Optional;
