+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/UserRepository.java
@@ -1,6 +1,6 @@
-package org.pallete.easelsocialservice.persistence;
+package org.palette.easelsocialservice.persistence;
 
-import org.pallete.easelsocialservice.persistence.domain.User;
+import org.palette.easelsocialservice.persistence.domain.User;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 public interface UserRepository extends Neo4jRepository<User, Long> {
