+++ src/social-service/src/main/java/org/palette/easelsocialservice/persistence/domain/Creates.java
@@ -1,4 +1,4 @@
-package org.pallete.easelsocialservice.persistence.domain;
+package org.palette.easelsocialservice.persistence.domain;
 
 import lombok.Getter;
 import org.springframework.data.neo4j.core.schema.RelationshipId;
