+++ src/social-service/src/main/java/org/palette/easelsocialservice/service/PaintService.java
@@ -1,0 +1,4 @@
+package org.palette.easelsocialservice.service;
+
+public class PaintService {
+}
