--- src/social-service/src/main/java/org/pallete/easelsocialservice/service/PaintService.java
@@ -1,4 +1,0 @@
-package org.pallete.easelsocialservice.service;
-
-public class PaintService {
-}
